extract HistoryInserter to another class
